+++ src/Http/ClientException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Ilovepdf\Http;
+
+class ClientException extends \GuzzleHttp\Exception\ClientException
+{
+}