+++ changelog.md
@@ -1,6 +1,26 @@
 Changelog
 ---------
 
+
+v1.2.2
+
+* Compatibility with php 8.1
+* Upgraded php version requirements
+* Only composer use mode
+* Code cleaned
+* Improved exception messages
+
+v1.2.1
+
+* New tool editpdf added
+
+v1.1.18
+
+* Php 8 compatibility
+* Upgraded php version requirements
+* Added Watermark with images
+* Improved download to browser method
+
 v1.0.11
 
 * Bug AuthException warning